requirements.txt
Flask==3.0.2
Flask-Caching==2.1.0
Flask-CORS==5.0.1
Flask-RESTX==1.3.0
python-dotenv==1.0.1
requests==2.31.0
beautifulsoup4==4.12.3
Werkzeug==3.0.1
gunicorn==21.2.0
APScheduler==3.11.0
openai==1.12.0
SQLAlchemy==2.0.41